backend/requirements.txt
# Backend Dependencies for PII Proxy Architecture

# LiteLLM - Unified interface for LLM APIs
litellm>=1.35.0

# Microsoft Presidio - PII detection and anonymization
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.6.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0-py3-none-any.whl

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Authentication and Security
PyJWT>=2.8.0
cryptography>=41.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Caching and Performance
redis>=4.6.0
cachetools>=5.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0

# Logging and Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0